requirements.txt
django==5.0.4
django-admin-interface
django-crispy-forms
crispy-bootstrap4
crispy-bootstrap5
pillow
zeep
psycopg2-binary
django-admin-confirm
django-axes
django-simple-captcha
django-recaptcha
cloudinary
whitenoise
djangorestframework
django-cloudinary-storage

# pip install -r requirements.txt